+++ Gemini200SLogAnalyzer/Models/LogDataSourceType.cs
@@ -1,0 +1,7 @@
+namespace Gemini200SLogAnalyzer.Models;
+
+public enum LogDataSourceType
+{
+    WaferLog,
+    ManualLog
+}